% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasu.R
\name{sasu}
\alias{sasu}
\title{Semi-Analytical Spin-Up}
\usage{
sasu(iterations, timestep, Initial_X, A, B, K, Tr, Scalar, Bscalar, Gpp)
}
\arguments{
\item{iterations}{number of iterations of the model}

\item{timestep}{a character representing the time interval between each sample; timestep = "month" or "day"}

\item{Initial_X}{a vector of the initial size of each C pool}

\item{A}{a matrix indicating coefficients of transfer among C pools}

\item{B}{a vector indicating indicating plant C allocation; Only plant C pools will not have a value of 0}

\item{K}{a diagonal matrix indicating C leaving individual pools through mortality or decomposition}

\item{Tr}{a matrix indicating vertical transfer coefficients}

\item{Scalar}{a matrix indicating environmental scalars, which is used to modify the matrix K}

\item{Bscalar}{a matrix indicating C input through photosynthesis, which is used to modify the matrix B}

\item{Gpp}{a vector of daily or monthly gross primary production(Gpp)}
}
\value{
a vector of the initial size of each C pool
}
\description{
A semi-analytical solution to accelerate spin-up
}
\examples{
asdsadasd
asdasdasd
}
