% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_run.R
\name{forward_run}
\alias{forward_run}
\title{Semi-Analytical Spin-Up}
\usage{
forward_run(iterations, timestep, Initial_X, A, B, K, Tr, Scalar, Bscalar, Gpp)
}
\description{
This package predictswhether the stock price at tommorow's market close would be higher or lowercompared to today's closing place.
}
\examples{
sasu("month", x0, mat.A, mat.B0, mat.K, mat.Tr, scalar1, Bscalar1, I4simu0)
}
