% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_run.R
\name{forward_run}
\alias{forward_run}
\title{Forward Run}
\usage{
forward_run(Initial_X, A, B, K, Tr, Scalar, Bscalar, Gpp)
}
\arguments{
\item{Initial_X}{a vector of the initial size of each C pool, which can be obtained by "sasu"}
}
\value{
a list
}
\description{
predict C pool sizes and C fluxes
}
\examples{
sasu("month", x0, mat.A, mat.B0, mat.K, mat.Tr, scalar1, Bscalar1, I4simu0)
}
